llm-integration/py-api/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
oracledb
httpx
openai
sqlalchemy-orm
async-sqlalchemy
aiosqlite
chromadb